Added in custom multipart request, which uses classes from Apache HttpClient, which unfortunately conflicts with Android's HttpClient (which by the way doesn't support multpart forms), which is a problem because we have to selectively include jars from Apache HttpClient. This works totally fine in a new Android project that has no other dependencies BUT omygodwhy Realm doesn't play along nicely and is unable to run its annotation processor to generate necessary proxy classes.
Easiest solution is to use another library for multipart form requests, such as OkHttp, but this totally breaks the slick API queue structure we have going on. Other solution is to write the multipart code ourselves, which might not actually be that terrible if we just copy what Apache is doing.